Clean up the GraphLoaderCache and move logic from the models into that utility
