+++ Sources/TuistCore/Graph/GraphCircularDetector.swift
